third_party/WebKit/LayoutTests/http/tests/security/suborigins/crossorigin/suborigin-cross-origin-script-window-onerror-cors-expected.txt

CONSOLE ERROR: line 1: Uncaught SomeError
This is a testharness.js-based test.
PASS Verify that thrown exception is unsanitized when crossorigin script loaded with crossorigin attribute
Harness: the test ran to completion.
